client/requirements.txt
transformers==4.45.0
torch>=2.0.0
requests
pyyaml
tenseal
